move add_prefix_suffix, autocorr, transpose
