+++ python/cudf/cudf/tests/test_series.py
@@ -72,71 +72,6 @@ def test_series_error_equality(sr1, sr2, op):
     assert_exceptions_equal(op, op, ([sr1, sr2],), ([gsr1, gsr2],))
 
 
-def test_series_add_prefix():
-    cd_s = cudf.Series([1, 2, 3, 4])
-    pd_s = cd_s.to_pandas()
-
-    got = cd_s.add_prefix("item_")
-    expected = pd_s.add_prefix("item_")
-
-    assert_eq(got, expected)
-
-
-def test_series_add_suffix():
-    cd_s = cudf.Series([1, 2, 3, 4])
-    pd_s = cd_s.to_pandas()
-
-    got = cd_s.add_suffix("_item")
-    expected = pd_s.add_suffix("_item")
-
-    assert_eq(got, expected)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        [0.25, 0.5, 0.2, -0.05],
-        [0, 1, 2, np.nan, 4, cudf.NA, 6],
-    ],
-)
-@pytest.mark.parametrize("lag", [1, 2, 3, 4])
-def test_autocorr(data, lag):
-    cudf_series = cudf.Series(data)
-    psr = cudf_series.to_pandas()
-
-    cudf_corr = cudf_series.autocorr(lag=lag)
-
-    # autocorrelation is undefined (nan) for less than two entries, but pandas
-    # short-circuits when there are 0 entries and bypasses the numpy function
-    # call that generates an error.
-    num_both_valid = (psr.notna() & psr.shift(lag).notna()).sum()
-    with expect_warning_if(num_both_valid == 1, RuntimeWarning):
-        pd_corr = psr.autocorr(lag=lag)
-
-    assert_eq(pd_corr, cudf_corr)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        [0, 1, 2, 3],
-        ["abc", "a", None, "hello world", "foo buzz", "", None, "rapids ai"],
-    ],
-)
-def test_series_transpose(data):
-    psr = pd.Series(data=data)
-    csr = cudf.Series(data=data)
-
-    cudf_transposed = csr.transpose()
-    pd_transposed = psr.transpose()
-    cudf_property = csr.T
-    pd_property = psr.T
-
-    assert_eq(pd_transposed, cudf_transposed)
-    assert_eq(pd_property, cudf_property)
-    assert_eq(cudf_transposed, csr)
-
-
 @pytest.mark.parametrize(
     "data",
     [1, 3, 5, 7, 7],
